requirements.txt
streamlit==1.28.1
python-docx==1.1.0
langchain==0.0.350
langchain-groq==0.0.1
groq==0.4.2
scikit-learn==1.7.1
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0
tiktoken==0.5.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
